pyproject.toml
[project]
name = "wp-mpm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "h5py==3.9.0",
    "imageio>=2.37.2",
    "imageio-ffmpeg>=0.6.0",
    "numpy==1.26.4",
    "open3d>=0.19.0",
    "pillow>=12.0.0",
    "torch>=2.9.0",
    "torchaudio>=2.9.0",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "warp-lang>=1.10.1",
]
[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }
torchaudio = { index = "pytorch-cu128" }
